+++ gtfs-structure-2/src/lib.rs
@@ -1,5 +1,5 @@
 #![feature(slice_group_by)]
-pub type IdType = (u8, u64);
+pub type IdType = (u16, u64);
 
 pub mod calendar;
 pub mod gtfs_wrapper;
